added a local storage login sys

+++ src/firebase/deleteItem.ts
@@ -2,18 +2,21 @@ import { firestore } from "./config";
 
 const deleteItem = (
   id: string,
-  setEve: React.Dispatch<React.SetStateAction<boolean>>
+  setEve: React.Dispatch<React.SetStateAction<boolean>>,
+  collectionName: string
 ) => {
-  firestore
-    .collection("cart")
-    .doc(id)
-    .delete()
-    .then(() => {
-      setEve(false);
-    })
-    .catch((err) => {
-      console.error("errorrr", err);
-    });
+  if (collectionName !== "") {
+    firestore
+      .collection(collectionName)
+      .doc(id)
+      .delete()
+      .then(() => {
+        setEve(false);
+      })
+      .catch((err) => {
+        console.error("errorrr", err);
+      });
+  }
 };
 
 export default deleteItem;
